Step2 (#29)
* chore: 테스트 함수명 given-when-then 순서로 변경

* refactor: MVP 패턴으로 변경

* test: test 함수에서 사용하는 변수 매번 초기화 되도록 수정

* chore: operand 와 호환을 위해 x를 *로 수정

* test: Presenter 단위 테스트 작성

* test: mock stubbing

* chore: operand 와 호환을 위해 ÷을 /로 수정

* fix: 소수로 인한 NumberFormatException 발생 처리

+++ domain/src/main/java/com/github/dodobest/domain/InputHandler.kt
@@ -19,7 +19,7 @@ class InputHandler {
 
     fun handleInputNum (inputNum: String) {
         if (arraySize > 0 && expression.isNum(numAndSignArray.last().toCharArray().last())) {
-            numAndSignArray[arraySize-1] = (numAndSignArray.last().toInt() * 10 + inputNum.toInt()).toString()
+            numAndSignArray[arraySize-1] = (numAndSignArray.last().toDouble().toInt() * 10 + inputNum.toInt()).toString()
             return
         }
         numAndSignArray.add(inputNum)
